feat!: checkers are ALMOST DONE
- changed idx from int to size_t
- finished checker for KICK, PRIVMSG, PART
- renamed ::chans_::hasChanAuthorisations to isOp

+++ headers/builder/command-validation/CmdSpec.hpp
@@ -77,7 +77,7 @@ class CmdSpec {
 		~CmdBuilder(void);
 
 		//methods
-		CmdBuilder &addChecker(bool (*ft)(CmdSpec &cmd, int idx));
+		CmdBuilder &addChecker(bool (*ft)(CmdSpec &cmd, size_t idx));
 		CmdBuilder &CmExecutor(void (*ft)(CmdSpec &cmd));
 		CmdBuilder &Name(const std::string &name);
 		CmdBuilder &addParam(e_param type, CmdParam *param);
@@ -88,7 +88,7 @@ class CmdSpec {
 		int registrationStage_;
 		paramMap params_;
 		std::string name_;
-		std::vector< bool (*)(CmdSpec &cmd, int idx) > checkers_;
+		std::vector< bool (*)(CmdSpec &cmd, size_t idx) > checkers_;
 		void (*cmExecutor_)(CmdSpec &cmd);
 	};
 
@@ -99,12 +99,12 @@ class CmdSpec {
 	const std::string name_;
 	int registrationStage_;
 	paramMap params_;
-	std::vector< bool (*)(CmdSpec &cmd, int idx) > checkers_;
+	std::vector< bool (*)(CmdSpec &cmd, size_t idx) > checkers_;
 	void (*cmExecutor_)(CmdSpec &cmd);
 
 	// private constructor
 	CmdSpec(const std::string name, int registrationStage, paramMap params,
-			std::vector< bool (*)(CmdSpec &, int) > checkers,
+			std::vector< bool (*)(CmdSpec &, size_t) > checkers,
 			void (*cmExecutor)(CmdSpec &cmd));
 };
